brv-app/requirements.txt
streamlit==1.22.0
pandas==1.5.3
numpy==1.24.4
qrcode==7.4.2
Pillow==9.5.0
pyinstaller==5.9.0
gspread==5.7.2
oauth2client==4.1.3
pyyaml==6.0
requests==2.28.2
flask==2.3.3
python-dotenv==1.0.0
google-api-python-client==2.97.0
# Oracle Cloud dependencies
oracledb==1.4.1
bcrypt==4.0.1
# Added dependencies for Docker
pymysql
sqlalchemy
google-auth
google-auth-httplib2
google-auth-oauthlib
# Removed dependencies (commented out for reference)
# mysql-connector-python==8.0.33